requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cssutils==2.2.0
email-to==0.1.0
Flask==1.1.2
future==0.18.2
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mime==0.1.0
premailer==3.8.0
requests==2.25.1
secure-smtplib==0.1.1
SimpleTelnetMail==0.1.1
soupsieve==2.2.1
urllib3==1.26.4
Werkzeug==1.0.1
dnspython
pymongo
flask-pymongo
wtforms
flask-mail
1337x